Consolidate CDP client transport
