+++ server/lib/cdpclient/cdpclient.go
@@ -3,10 +3,12 @@ package cdpclient
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
 	"strings"
+	"sync"
 	"sync/atomic"
 	"time"
 
@@ -20,26 +22,52 @@ type cdpRequest struct {
 	SessionID string          `json:"sessionId,omitempty"`
 }
 
-type cdpResponse struct {
-	ID     int64           `json:"id"`
-	Result json.RawMessage `json:"result,omitempty"`
-	Error  *cdpError       `json:"error,omitempty"`
+// ErrOutcomeUnknown means a CDP connection closed while a command was in
+// flight, so the caller cannot know whether Chromium applied it.
+var ErrOutcomeUnknown = errors.New("CDP command outcome is unknown")
+
+// Message is a response or event received from Chromium.
+type Message struct {
+	ID        int64           `json:"id,omitempty"`
+	Method    string          `json:"method,omitempty"`
+	Params    json.RawMessage `json:"params,omitempty"`
+	Result    json.RawMessage `json:"result,omitempty"`
+	Error     *Error          `json:"error,omitempty"`
+	SessionID string          `json:"sessionId,omitempty"`
 }
 
-type cdpError struct {
+// Error is a protocol error returned by Chromium.
+type Error struct {
 	Code    int    `json:"code"`
 	Message string `json:"message"`
 }
 
-func (e *cdpError) Error() string {
+func (e *Error) Error() string {
 	return fmt.Sprintf("CDP error %d: %s", e.Code, e.Message)
 }
 
-// Client is a minimal CDP client that communicates over a browser-level
-// DevTools WebSocket connection.
+type commandResult struct {
+	result json.RawMessage
+	err    error
+}
+
+// Client maintains one browser-level DevTools connection. Commands may be
+// sent concurrently. Clients created by DialWithEvents also expose protocol
+// events through Events.
 type Client struct {
 	conn   *websocket.Conn
-	nextID atomic.Int64
+	ctx    context.Context
+	cancel context.CancelFunc
+
+	nextID  atomic.Int64
+	writeMu sync.Mutex
+
+	pendingMu sync.Mutex
+	pending   map[int64]chan commandResult
+	events    chan Message
+
+	closed    chan struct{}
+	closeOnce sync.Once
 }
 
 // BrowserWebSocketURL reads Chrome's browser-level DevTools WebSocket URL.
@@ -68,25 +96,60 @@ func BrowserWebSocketURL(ctx context.Context, versionURL string) (string, error)
 	return version.WebSocketDebuggerURL, nil
 }
 
-// Dial opens a WebSocket connection to the given DevTools URL.
+// Dial opens a command-only DevTools connection. Protocol events are
+// discarded, preserving the behavior expected by existing short-lived users.
 func Dial(ctx context.Context, devtoolsURL string) (*Client, error) {
+	return dial(ctx, devtoolsURL, false)
+}
+
+// DialWithEvents opens a DevTools connection that delivers protocol events in
+// receive order through Events.
+func DialWithEvents(ctx context.Context, devtoolsURL string) (*Client, error) {
+	return dial(ctx, devtoolsURL, true)
+}
+
+func dial(ctx context.Context, devtoolsURL string, withEvents bool) (*Client, error) {
 	conn, _, err := websocket.Dial(ctx, devtoolsURL, nil)
 	if err != nil {
 		return nil, fmt.Errorf("dial devtools: %w", err)
 	}
-	conn.SetReadLimit(4 * 1024 * 1024)
-	return &Client{conn: conn}, nil
+	readLimit := int64(4 * 1024 * 1024)
+	var events chan Message
+	if withEvents {
+		readLimit = 8 * 1024 * 1024
+		events = make(chan Message, 256)
+	}
+	conn.SetReadLimit(readLimit)
+	clientCtx, cancel := context.WithCancel(context.Background())
+	client := &Client{
+		conn:    conn,
+		ctx:     clientCtx,
+		cancel:  cancel,
+		pending: make(map[int64]chan commandResult),
+		events:  events,
+		closed:  make(chan struct{}),
+	}
+	go client.readLoop()
+	return client, nil
 }
 
-// Close shuts down the WebSocket connection.
+// Close shuts down the WebSocket connection and unblocks pending commands.
 func (c *Client) Close() error {
-	return c.conn.Close(websocket.StatusNormalClosure, "done")
+	c.shutdown()
+	c.conn.CloseNow()
+	return nil
 }
 
-// send sends a CDP command and waits for the matching response, discarding
-// any intermediate events. This is safe for short-lived connections where the
-// caller controls the full message sequence.
-func (c *Client) send(ctx context.Context, method string, params any, sessionID string) (json.RawMessage, error) {
+func (c *Client) shutdown() {
+	c.closeOnce.Do(func() {
+		c.cancel()
+		close(c.closed)
+		c.failPending(ErrOutcomeUnknown)
+	})
+}
+
+// Send sends a CDP command and waits for its matching response.
+func (c *Client) Send(ctx context.Context, method string, params any, sessionID string) (json.RawMessage, error) {
 	id := c.nextID.Add(1)
 
 	var rawParams json.RawMessage
@@ -98,33 +161,117 @@ func (c *Client) send(ctx context.Context, method string, params any, sessionID 
 		rawParams = b
 	}
 
-	req := cdpRequest{ID: id, Method: method, Params: rawParams, SessionID: sessionID}
-	reqBytes, err := json.Marshal(req)
+	reqBytes, err := json.Marshal(cdpRequest{ID: id, Method: method, Params: rawParams, SessionID: sessionID})
 	if err != nil {
 		return nil, fmt.Errorf("marshal request: %w", err)
 	}
 
-	if err := c.conn.Write(ctx, websocket.MessageText, reqBytes); err != nil {
+	responseCh := make(chan commandResult, 1)
+	c.pendingMu.Lock()
+	c.pending[id] = responseCh
+	c.pendingMu.Unlock()
+
+	c.writeMu.Lock()
+	err = c.conn.Write(ctx, websocket.MessageText, reqBytes)
+	c.writeMu.Unlock()
+	if err != nil {
+		c.pendingMu.Lock()
+		delete(c.pending, id)
+		c.pendingMu.Unlock()
 		return nil, fmt.Errorf("write: %w", err)
 	}
 
+	select {
+	case response := <-responseCh:
+		return response.result, response.err
+	case <-ctx.Done():
+		c.pendingMu.Lock()
+		delete(c.pending, id)
+		c.pendingMu.Unlock()
+		select {
+		case response := <-responseCh:
+			return response.result, response.err
+		default:
+			return nil, fmt.Errorf("read: %w", ctx.Err())
+		}
+	case <-c.closed:
+		select {
+		case response := <-responseCh:
+			return response.result, response.err
+		default:
+			return nil, fmt.Errorf("read: %w", ErrOutcomeUnknown)
+		}
+	}
+}
+
+// Events returns the event stream for clients created by DialWithEvents. It
+// returns nil for command-only clients.
+func (c *Client) Events() <-chan Message {
+	return c.events
+}
+
+// Done closes when the connection shuts down.
+func (c *Client) Done() <-chan struct{} {
+	return c.closed
+}
+
+// IsClosed reports whether the connection has shut down.
+func (c *Client) IsClosed() bool {
+	select {
+	case <-c.closed:
+		return true
+	default:
+		return false
+	}
+}
+
+func (c *Client) readLoop() {
+	if c.events != nil {
+		defer close(c.events)
+	}
 	for {
-		_, msg, err := c.conn.Read(ctx)
+		_, payload, err := c.conn.Read(c.ctx)
 		if err != nil {
-			return nil, fmt.Errorf("read: %w", err)
+			c.shutdown()
+			return
+		}
+		var message Message
+		if json.Unmarshal(payload, &message) != nil {
+			continue
+		}
+		if message.ID == 0 {
+			if c.events == nil {
+				continue
+			}
+			select {
+			case c.events <- message:
+			case <-c.closed:
+				return
+			}
+			continue
 		}
 
-		var resp cdpResponse
-		if err := json.Unmarshal(msg, &resp); err != nil {
-			continue // skip malformed messages
+		c.pendingMu.Lock()
+		responseCh := c.pending[message.ID]
+		delete(c.pending, message.ID)
+		c.pendingMu.Unlock()
+		if responseCh == nil {
+			continue
 		}
-		if resp.ID != id {
-			continue // skip events and responses to other commands
+		if message.Error != nil {
+			responseCh <- commandResult{err: message.Error}
+		} else {
+			responseCh <- commandResult{result: message.Result}
 		}
-		if resp.Error != nil {
-			return nil, resp.Error
-		}
-		return resp.Result, nil
+	}
+}
+
+func (c *Client) failPending(err error) {
+	c.pendingMu.Lock()
+	defer c.pendingMu.Unlock()
+	for id, responseCh := range c.pending {
+		responseCh <- commandResult{err: err}
+		delete(c.pending, id)
 	}
 }
 
@@ -151,7 +298,7 @@ type BrowserVersion struct {
 // listener of a Chromium that has not yet wired up its CDP routes. A
 // Browser.getVersion round-trip rules out both cases.
 func (c *Client) GetBrowserVersion(ctx context.Context) (*BrowserVersion, error) {
-	raw, err := c.send(ctx, "Browser.getVersion", nil, "")
+	raw, err := c.Send(ctx, "Browser.getVersion", nil, "")
 	if err != nil {
 		return nil, fmt.Errorf("Browser.getVersion: %w", err)
 	}
@@ -165,7 +312,7 @@ func (c *Client) GetBrowserVersion(ctx context.Context) (*BrowserVersion, error)
 // LoadUnpackedExtension installs an unpacked extension from an absolute path
 // visible to Chromium and returns its extension ID.
 func (c *Client) LoadUnpackedExtension(ctx context.Context, path string) (string, error) {
-	raw, err := c.send(ctx, "Extensions.loadUnpacked", map[string]string{"path": path}, "")
+	raw, err := c.Send(ctx, "Extensions.loadUnpacked", map[string]string{"path": path}, "")
 	if err != nil {
 		return "", fmt.Errorf("Extensions.loadUnpacked: %w", err)
 	}
@@ -192,7 +339,7 @@ type ExtensionInfo struct {
 
 // GetExtensions returns all unpacked extensions known to Chromium.
 func (c *Client) GetExtensions(ctx context.Context) ([]ExtensionInfo, error) {
-	raw, err := c.send(ctx, "Extensions.getExtensions", nil, "")
+	raw, err := c.Send(ctx, "Extensions.getExtensions", nil, "")
 	if err != nil {
 		return nil, fmt.Errorf("Extensions.getExtensions: %w", err)
 	}
@@ -227,7 +374,7 @@ type HistogramBucket struct {
 // reads Chrome's in-memory UMA histograms without attaching to any page.
 // query is a substring filter on the histogram name; empty returns all.
 func (c *Client) GetHistograms(ctx context.Context, query string) ([]Histogram, error) {
-	raw, err := c.send(ctx, "Browser.getHistograms", map[string]any{"query": query}, "")
+	raw, err := c.Send(ctx, "Browser.getHistograms", map[string]any{"query": query}, "")
 	if err != nil {
 		return nil, fmt.Errorf("Browser.getHistograms: %w", err)
 	}
@@ -242,7 +389,7 @@ func (c *Client) GetHistograms(ctx context.Context, query string) ([]Histogram, 
 
 // CountPageTargets returns the number of open page targets.
 func (c *Client) CountPageTargets(ctx context.Context) (int, error) {
-	targetsResult, err := c.send(ctx, "Target.getTargets", nil, "")
+	targetsResult, err := c.Send(ctx, "Target.getTargets", nil, "")
 	if err != nil {
 		return 0, fmt.Errorf("Target.getTargets: %w", err)
 	}
@@ -273,7 +420,7 @@ func DispatchStartURL(ctx context.Context, devtoolsURL, url string) error {
 	}
 	defer c.Close()
 
-	targetsResult, err := c.send(ctx, "Target.getTargets", nil, "")
+	targetsResult, err := c.Send(ctx, "Target.getTargets", nil, "")
 	if err != nil {
 		return fmt.Errorf("Target.getTargets: %w", err)
 	}
@@ -297,12 +444,12 @@ func DispatchStartURL(ctx context.Context, devtoolsURL, url string) error {
 			pageTargetID = t.TargetID
 			continue
 		}
-		_, _ = c.send(ctx, "Target.closeTarget", map[string]any{
+		_, _ = c.Send(ctx, "Target.closeTarget", map[string]any{
 			"targetId": t.TargetID,
 		}, "")
 	}
 	if pageTargetID == "" {
-		createResult, err := c.send(ctx, "Target.createTarget", map[string]any{
+		createResult, err := c.Send(ctx, "Target.createTarget", map[string]any{
 			"url": "about:blank",
 		}, "")
 		if err != nil {
@@ -317,7 +464,7 @@ func DispatchStartURL(ctx context.Context, devtoolsURL, url string) error {
 		pageTargetID = created.TargetID
 	}
 
-	attachResult, err := c.send(ctx, "Target.attachToTarget", map[string]any{
+	attachResult, err := c.Send(ctx, "Target.attachToTarget", map[string]any{
 		"targetId": pageTargetID,
 		"flatten":  true,
 	}, "")
@@ -334,12 +481,12 @@ func DispatchStartURL(ctx context.Context, devtoolsURL, url string) error {
 	defer func() {
 		detachCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 		defer cancel()
-		_, _ = c.send(detachCtx, "Target.detachFromTarget", map[string]any{
+		_, _ = c.Send(detachCtx, "Target.detachFromTarget", map[string]any{
 			"sessionId": attach.SessionID,
 		}, "")
 	}()
 
-	if _, err := c.send(ctx, "Page.navigate", map[string]any{"url": url}, attach.SessionID); err != nil {
+	if _, err := c.Send(ctx, "Page.navigate", map[string]any{"url": url}, attach.SessionID); err != nil {
 		return fmt.Errorf("Page.navigate: %w", err)
 	}
 	return nil
@@ -366,7 +513,7 @@ func DispatchStartURLAndWait(ctx context.Context, devtoolsURL, navigationURL, de
 	if err != nil {
 		return err
 	}
-	attachResult, err := c.send(ctx, "Target.attachToTarget", map[string]any{
+	attachResult, err := c.Send(ctx, "Target.attachToTarget", map[string]any{
 		"targetId": targetID,
 		"flatten":  true,
 	}, "")
@@ -382,7 +529,7 @@ func DispatchStartURLAndWait(ctx context.Context, devtoolsURL, navigationURL, de
 	defer func() {
 		detachCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 		defer cancel()
-		_, _ = c.send(detachCtx, "Target.detachFromTarget", map[string]any{
+		_, _ = c.Send(detachCtx, "Target.detachFromTarget", map[string]any{
 			"sessionId": attach.SessionID,
 		}, "")
 	}()
@@ -395,7 +542,7 @@ func DispatchStartURLAndWait(ctx context.Context, devtoolsURL, navigationURL, de
 	}
 	lastNavigate := time.Now()
 	for {
-		raw, evalErr := c.send(ctx, "Runtime.evaluate", map[string]any{
+		raw, evalErr := c.Send(ctx, "Runtime.evaluate", map[string]any{
 			"expression":    `JSON.stringify({url: location.href, readyState: document.readyState})`,
 			"returnByValue": true,
 		}, attach.SessionID)
@@ -412,7 +559,7 @@ func DispatchStartURLAndWait(ctx context.Context, devtoolsURL, navigationURL, de
 					return nil
 				}
 				if strings.HasPrefix(lastState.URL, "chrome-error://") && time.Since(lastNavigate) >= 250*time.Millisecond {
-					_, _ = c.send(ctx, "Page.navigate", map[string]any{"url": navigationURL}, attach.SessionID)
+					_, _ = c.Send(ctx, "Page.navigate", map[string]any{"url": navigationURL}, attach.SessionID)
 					lastNavigate = time.Now()
 				}
 			}
@@ -436,7 +583,7 @@ func relatedHosts(a, b string) bool {
 // by Target.getTargets. Callers that need to operate on the user-facing
 // browser window (Emulation, Browser.* window bounds) use this to find it.
 func (c *Client) firstPageTargetID(ctx context.Context) (string, error) {
-	targetsResult, err := c.send(ctx, "Target.getTargets", nil, "")
+	targetsResult, err := c.Send(ctx, "Target.getTargets", nil, "")
 	if err != nil {
 		return "", fmt.Errorf("Target.getTargets: %w", err)
 	}
@@ -483,7 +630,7 @@ func (c *Client) SetWindowBoundsMaximized(ctx context.Context) error {
 		return nil
 	}
 
-	if _, err := c.send(ctx, "Browser.setWindowBounds", map[string]any{
+	if _, err := c.Send(ctx, "Browser.setWindowBounds", map[string]any{
 		"windowId": bounds.WindowID,
 		"bounds":   map[string]any{"windowState": "maximized"},
 	}, ""); err != nil {
@@ -512,7 +659,7 @@ func (c *Client) GetWindowBounds(ctx context.Context) (WindowBounds, error) {
 		return WindowBounds{}, err
 	}
 
-	winRaw, err := c.send(ctx, "Browser.getWindowForTarget", map[string]any{"targetId": pageTargetID}, "")
+	winRaw, err := c.Send(ctx, "Browser.getWindowForTarget", map[string]any{"targetId": pageTargetID}, "")
 	if err != nil {
 		return WindowBounds{}, fmt.Errorf("Browser.getWindowForTarget: %w", err)
 	}
@@ -544,7 +691,7 @@ func (c *Client) SetDeviceMetricsOverride(ctx context.Context, width, height int
 		return err
 	}
 
-	attachResult, err := c.send(ctx, "Target.attachToTarget", map[string]any{
+	attachResult, err := c.Send(ctx, "Target.attachToTarget", map[string]any{
 		"targetId": pageTargetID,
 		"flatten":  true,
 	}, "")
@@ -559,7 +706,7 @@ func (c *Client) SetDeviceMetricsOverride(ctx context.Context, width, height int
 		return fmt.Errorf("unmarshal attach: %w", err)
 	}
 
-	_, err = c.send(ctx, "Emulation.setDeviceMetricsOverride", map[string]any{
+	_, err = c.Send(ctx, "Emulation.setDeviceMetricsOverride", map[string]any{
 		"width":             width,
 		"height":            height,
 		"deviceScaleFactor": 1,
@@ -569,7 +716,7 @@ func (c *Client) SetDeviceMetricsOverride(ctx context.Context, width, height int
 		return fmt.Errorf("Emulation.setDeviceMetricsOverride: %w", err)
 	}
 
-	_, _ = c.send(ctx, "Target.detachFromTarget", map[string]any{
+	_, _ = c.Send(ctx, "Target.detachFromTarget", map[string]any{
 		"sessionId": attach.SessionID,
 	}, "")
